refactor(config): merged file
